David's issue and fix neighbors / inverted neighbors count

+++ src/modules/AutoMoDeConditionNeighborsCount.cpp
@@ -49,7 +49,7 @@
 
 	bool AutoMoDeConditionNeighborsCount::Verify() {
 		UInt32 unNumberNeighbors = m_pcRobotDAO->GetNumberNeighbors();
-		Real fProbability = (1/(1 + exp(m_fParameterEta * (m_unParameterXi - unNumberNeighbors))));
+                Real fProbability = (1/(1 + exp(m_fParameterEta * ((int)m_unParameterXi - (int)unNumberNeighbors))));
 		return EvaluateBernoulliProbability(fProbability);
 	}
